backend/requirements.txt
Django==5.1.1
djangorestframework==3.14.0
django-cors-headers==4.6.0
django-environ==0.10.0
requests==2.31.0
gunicorn==20.1.0

